requirements.txt
openai>=1.0.0
pandas>=1.3.0
networkx>=2.6.0
ipycytoscape>=1.3.1
ipywidgets>=7.6.0
python-dotenv>=0.19.0
anthropic>=0.50.0 